Minor optimization fixes

+++ Engine/Input/InputManager.cpp
@@ -3,15 +3,15 @@
 
 void InputManager::HandleInput(int key, int action)
 {
-	if (key < 0 || InputKey::LAST_KEY_VALUE < key || action < 0 || InputAction::LAST_ACTION_VALUE < action)
+	if (key < 0 || (int)InputKey::LAST_KEY_VALUE < key || action < 0 || (int)InputAction::LAST_ACTION_VALUE < action)
 	{ 
 		return;
 	}
-	if (BindedActions.find(action * InputKey::LAST_KEY_VALUE + key) == BindedActions.end())
+	if (BindedActions.find(action * (int)InputKey::LAST_KEY_VALUE + key) == BindedActions.end())
 	{
 		return;
 	}
-	for (std::function<void()> Action : BindedActions[action * InputKey::LAST_KEY_VALUE + key])
+	for (std::function<void()> Action : BindedActions[action * (int)InputKey::LAST_KEY_VALUE + key])
 	{
 		std::invoke(Action);
 	}
@@ -19,12 +19,12 @@ void InputManager::HandleInput(int key, int action)
 
 void InputManager::BindInput(InputActionFunction func, InputKey key, InputAction action)
 {
-	BindedActions[action * InputKey::LAST_KEY_VALUE + key].emplace_back(func);
+	BindedActions[(int)action * (int)InputKey::LAST_KEY_VALUE + (int)key].emplace_back(func);
 }
 
 void InputManager::BindInput(InputActionFunction func, int key, int action)
 {
-	if (key >= 0 && InputKey::LAST_KEY_VALUE >= key && action >= 0 && InputAction::LAST_ACTION_VALUE >= action)
+	if (key >= 0 && (int)InputKey::LAST_KEY_VALUE >= key && action >= 0 && (int)InputAction::LAST_ACTION_VALUE >= action)
 	{
 		BindInput(func, static_cast<InputKey>(key), static_cast<InputAction>(action));
 	}
